% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpression_sim.R
\name{BranchCal}
\alias{BranchCal}
\title{Simulate observed count matrix given technical biases and the true counts}
\usage{
BranchCal(phyla, cell_table)
}
\arguments{
\item{phyla}{cell state tree}

\item{cell_table}{a dataframe giving the number of cells per cell type}
}
\description{
Simulate observed count matrix given technical biases and the true counts
}
