% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpression_sim.R
\name{GeneEffects}
\alias{GeneEffects}
\title{Getting GeneEffects matrices}
\usage{
GeneEffects(ngenes, ncif, randseed, prob, geffect_mean, geffect_sd)
}
\arguments{
\item{ngenes}{number of genes}

\item{ncif}{number of cifs}

\item{prob}{the probability that the effect size is not 0}

\item{geffect_mean}{the mean of the normal distribution where the non-zero effect sizes are dropped from}

\item{geffect_sd}{the standard deviation of the normal distribution where the non-zero effect sizes are dropped from}

\item{randomseed}{(should produce same result if ngenes, ncif and randseed are all the same)}
}
\value{
a list of 3 matrices, each of dimension ngenes * ncif
}
\description{
This function randomly generates the effect size of each cif on the dynamic expression parameters
}
