% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageBarcode_sim.R
\name{generate_dropout}
\alias{generate_dropout}
\title{Generate excision dropout}
\usage{
generate_dropout(barcode, mutation_cites)
}
\arguments{
\item{barcode}{input barcode with mutations}

\item{mutation_cites}{the mutated target sites}
}
\description{
Generate excision dropout
}
