% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageBarcode_sim.R
\name{generate_mutation}
\alias{generate_mutation}
\title{Generate CRISPR induced mutations}
\usage{
generate_mutation(
  barcode,
  mu = 0.1,
  N_ms = 100,
  mutation_dist = NULL,
  p_d = 0,
  unif_on = FALSE
)
}
\arguments{
\item{barcode}{input barcode (from the parent cell)}

\item{mu}{mutation rate per target, default is 1}

\item{N_ms}{the number of mutated states required}

\item{mutation_dist}{input distribution for the mutated states}

\item{p_d}{whether or not to simulate dropout effects, 0 or 1}

\item{unif_on}{distribution mode of mutated states. TRUE: uniform distribution; FALSE: real fitted distribution}
}
\description{
Generate CRISPR induced mutations
}
