% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE_LB_sim.R
\name{Generate_profile_multi}
\alias{Generate_profile_multi}
\title{Generate joint profile of gene expression and lineage barcode for output}
\usage{
Generate_profile_multi(observed_counts, muts, states)
}
\arguments{
\item{observed_counts}{Observed gene expression matrix}

\item{muts}{lineage character matrix}

\item{states}{cell states}
}
\description{
Generate joint profile of gene expression and lineage barcode for output
}
