% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageBarcode_sim.R
\name{Mutated_state_dist}
\alias{Mutated_state_dist}
\title{Generate mutated states from a real dataset}
\usage{
Mutated_state_dist(N_ms, cm)
}
\arguments{
\item{N_ms}{the number of mutated states required}

\item{cm}{character matrix of the real dataset}
}
\description{
Generate mutated states from a real dataset
}
