% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PlotUmap}
\alias{PlotUmap}
\title{Perform UMAP}
\usage{
PlotUmap(
  meta,
  data,
  pca = T,
  n_pc,
  min_dist = 0.5,
  n_neighbors = 30,
  label,
  saving = F,
  plotname,
  system.color = T
)
}
\arguments{
\item{meta}{simulation parameters}

\item{data}{expression matrix}

\item{min_dist}{minimum distance to identify neighborhood,default is 0.5}

\item{n_neighbors}{The size of local neighborhood for manifold learning, default is 30}

\item{label}{the column name of the meta data that the points needs to be colored by}

\item{plotname}{the name of the jpeg file}
}
\description{
This function takes the ground truth (meta), the expresssion matrix (data), and performs UMAP, saves the result to a jpeg file with user specified name
}
