% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpression_sim.R
\name{SampleCIF}
\alias{SampleCIF}
\title{Generate Cell Identity Factors recursively via Brownian motion}
\usage{
SampleCIF(par, depth, anc_state, edges)
}
\arguments{
\item{par}{current parent node label}

\item{depth}{current depth of the parent node}

\item{anc_state}{accumulated state identity factor value}

\item{edges}{edge list of the cell lineage}
}
\description{
Generate Cell Identity Factors recursively via Brownian motion
}
