% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE_LB_sim.R
\name{SampleEdgeNew}
\alias{SampleEdgeNew}
\title{Generate child CIF based on parent CIF}
\usage{
SampleEdgeNew(
  edge,
  depth,
  anc_state,
  edges,
  sif_mean = NULL,
  S = NULL,
  cif = NULL,
  mu = 0.1,
  p_d = 0,
  barcode = NULL,
  flag = 0,
  N_ms = NULL,
  unif_on = FALSE,
  lambda = NULL
)
}
\arguments{
\item{edge}{current edge on the cell lineage tree}

\item{depth}{current depth of the parent node}

\item{anc_state}{Cell Identity Factor value of the parent node}

\item{edges}{edge list of the cell lineage tree}

\item{sif_mean}{state identity factor values}

\item{S}{state and state identity factor of the parent}

\item{cif}{cell identity factor matrix}

\item{mu}{mutation rate}

\item{p_d}{whether or not to have dropouts}

\item{barcode}{current barcode of the parent node}

\item{flag}{whether or not generate barcode along with gene expressions}

\item{N_ms}{Number of possible mutated states}

\item{unif_on}{if unif_on is TRUE, the mutated states will be synthetically generated using uniform distribution; otherwise, it will be sampled from a real dataset}

\item{lambda}{a num vector that indicates the value of lambda that weights the additional random walk value for different depths}
}
\description{
Generate child CIF based on parent CIF
}
