% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE_LB_sim.R
\name{Samplelineage}
\alias{Samplelineage}
\title{Sample Cell Identity Factors via Brownian motion, based on the cell lineage}
\usage{
Samplelineage(
  par,
  depth,
  anc_state,
  edges,
  muts = NULL,
  p_d = 0.1,
  mu = 0.1,
  edges_state,
  sif_mean = NULL,
  S = NULL,
  p_a = 0.8,
  cif = NULL,
  flag = NULL,
  barcode = NULL,
  N_ms = NULL,
  unif_on = FALSE
)
}
\arguments{
\item{par}{current parent node label}

\item{depth}{current depth of the parent node}

\item{anc_state}{Cell Identity Factor value of the parent node}

\item{edges}{edge list of the cell lineage tree}

\item{muts}{the mutation barcode matrix}

\item{p_d}{whether or not to have dropouts}

\item{mu}{mutation rate}

\item{edges_state}{edge list of the state tree}

\item{sif_mean}{state identity factor values}

\item{S}{state and state identity factor of the parent}

\item{p_a}{asymmetric division rate}

\item{cif}{cell identity factor matrix}

\item{flag}{whether or not generate barcode along with gene expressions}

\item{barcode}{current barcode of the parent node}

\item{N_ms}{Number of possible mutated states}

\item{unif_on}{if unif_on is TRUE, the mutated states will be synthetically generated using uniform distribution; otherwise, it will be sampled from a real dataset}
}
\description{
Sample Cell Identity Factors via Brownian motion, based on the cell lineage
}
