% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpression_sim.R
\name{SIFGenerate}
\alias{SIFGenerate}
\title{Simulate State Identity Factor values given a state tree}
\usage{
SIFGenerate(phyla, n_diff, step = 1)
}
\arguments{
\item{phyla}{state tree structure}

\item{n_diff}{number of diff-CIF simulated}

\item{step}{stepsize of sampling intermediate states}
}
\description{
Simulate State Identity Factor values given a state tree
}
