% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateCIFs.R
\name{SimulateCIFs}
\alias{SimulateCIFs}
\title{Simulate Cell Identity Factor Matrix With Lineage Barcodes}
\usage{
SimulateCIFs(
  ncells,
  phyla,
  cif_center = 1,
  Sigma = 0.5,
  p_a = 0.8,
  p_edge = NULL,
  n_CIF,
  n_diff,
  step = 1,
  p_d = 0.1,
  mu = 0.1,
  N_char = 9,
  N_ms = 100,
  unif_on = FALSE,
  SIF_res = NULL,
  max_walk = 2,
  lambda = 0.05,
  T_cell = NULL
)
}
\arguments{
\item{ncells}{Number of Cells}

\item{phyla}{Cell state tree}

\item{cif_center}{Mean of CIFs, default is 1}

\item{Sigma}{Standard deviation of non-diff CIFs, difault is 0.5}

\item{p_a}{The asymmetric division rate, default is 0.8}

\item{p_edge}{The edge transition probability table, default is NULL so that the child edges are chosen with equal possibilities}

\item{n_CIF}{Total number of SIFs, both non-diff and diff combined}

\item{n_diff}{Number of diff SIFs}

\item{step}{Sampling stepsize of diff-SIF Brownian motion}

\item{p_d}{Dropout rate of CRISPR/Cas9 lineage barcodes}

\item{mu}{mutation rate of CRISPR/Cas9 lineage barcodes}

\item{N_char}{number of character sites for CRISPR/Cas9 lineage barcodes}

\item{N_ms}{number of mutated states for CRISPR/Cas9 lineage barcodes}

\item{unif_on}{sampling from synthetic uniform-distributed mutated states. When set FALSE, mutated states will be drawn from a real experimental dataset}

\item{SIF_res}{Optional input for State Identity Factors. If not, the function will generate SIF internally.}

\item{max_walk}{maximum walk distance of one asymmetric division on the cell state tree}

\item{lambda}{a num vector that indicates the max and the min value of lambda that weights the additional random walk value}

\item{T_cell}{optional, a cell division tree}
}
\description{
Simulate Cell Identity Factor Matrix With Lineage Barcodes
}
