% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageBarcode_sim.R
\name{CaptureDrop}
\alias{CaptureDrop}
\title{Simulate capture dropout based on observed counts and barcode data}
\usage{
CaptureDrop(observed_counts, muts)
}
\arguments{
\item{observed_counts}{Observed counts of gene expressions}

\item{muts}{Unprocessed mutation barcodes}
}
\description{
Simulate capture dropout based on observed counts and barcode data
}
